move concepts out of ASoA.h
